daydreams-api/image-search/requirements.txt
azure-cognitiveservices-search-imagesearch==1.0.0
Flask==1.0.2
gunicorn==19.9.0
